credits updated by generating image
